Prepared DXF folder for export
Reorganized the DXF folders in preparation for the second export for
Steve.

Moved the folders for the first export into a STEVE EXPORT 1 folder.
Reorganized the DXF files for the 2nd export based on material rather
than subcomponent. Steve cares primarily about material to group parts
made of the same material and doesn't care which subcomponent they are
for. I combined the drivetrain and chassis DXF files with the launcher
files. Also renamed folders that started with a '.' since UNIX-based
file systems hide those folders.